Fixed conversion warning from MSVC

+++ tests/runtime_tests/cli.cpp
@@ -6,8 +6,8 @@ using snatch::matchers::contains_substring;
 using arg_vector = snatch::small_vector<const char*, snatch::max_command_line_args>;
 
 TEST_CASE("parse arguments empty", "[cli]") {
-    const arg_vector args  = {"test"};
-    auto             input = snatch::cli::parse_arguments(args.size(), args.data());
+    const arg_vector args = {"test"};
+    auto input = snatch::cli::parse_arguments(static_cast<int>(args.size()), args.data());
 
     REQUIRE(input.has_value());
     CHECK(input->executable == "test"sv);
@@ -15,8 +15,8 @@ TEST_CASE("parse arguments empty", "[cli]") {
 };
 
 TEST_CASE("parse arguments empty .exe", "[cli]") {
-    const arg_vector args  = {"test.exe"};
-    auto             input = snatch::cli::parse_arguments(args.size(), args.data());
+    const arg_vector args = {"test.exe"};
+    auto input = snatch::cli::parse_arguments(static_cast<int>(args.size()), args.data());
 
     REQUIRE(input.has_value());
     CHECK(input->executable == "test"sv);
@@ -24,8 +24,8 @@ TEST_CASE("parse arguments empty .exe", "[cli]") {
 };
 
 TEST_CASE("parse arguments empty .something.exe", "[cli]") {
-    const arg_vector args  = {"test.something.exe"};
-    auto             input = snatch::cli::parse_arguments(args.size(), args.data());
+    const arg_vector args = {"test.something.exe"};
+    auto input = snatch::cli::parse_arguments(static_cast<int>(args.size()), args.data());
 
     REQUIRE(input.has_value());
     CHECK(input->executable == "test.something"sv);
@@ -33,8 +33,8 @@ TEST_CASE("parse arguments empty .something.exe", "[cli]") {
 };
 
 TEST_CASE("parse arguments help (long form)", "[cli]") {
-    const arg_vector args  = {"test", "--help"};
-    auto             input = snatch::cli::parse_arguments(args.size(), args.data());
+    const arg_vector args = {"test", "--help"};
+    auto input = snatch::cli::parse_arguments(static_cast<int>(args.size()), args.data());
 
     REQUIRE(input.has_value());
     CHECK(input->executable == "test"sv);
@@ -45,8 +45,8 @@ TEST_CASE("parse arguments help (long form)", "[cli]") {
 };
 
 TEST_CASE("parse arguments help (short form)", "[cli]") {
-    const arg_vector args  = {"test", "-h"};
-    auto             input = snatch::cli::parse_arguments(args.size(), args.data());
+    const arg_vector args = {"test", "-h"};
+    auto input = snatch::cli::parse_arguments(static_cast<int>(args.size()), args.data());
 
     REQUIRE(input.has_value());
     CHECK(input->executable == "test"sv);
@@ -57,15 +57,15 @@ TEST_CASE("parse arguments help (short form)", "[cli]") {
 };
 
 TEST_CASE("parse arguments help (duplicate)", "[cli]") {
-    const arg_vector args  = {"test", "--help", "--help"};
-    auto             input = snatch::cli::parse_arguments(args.size(), args.data());
+    const arg_vector args = {"test", "--help", "--help"};
+    auto input = snatch::cli::parse_arguments(static_cast<int>(args.size()), args.data());
 
     REQUIRE(!input.has_value());
 };
 
 TEST_CASE("parse arguments verbosity (long form)", "[cli]") {
-    const arg_vector args  = {"test", "--verbosity", "high"};
-    auto             input = snatch::cli::parse_arguments(args.size(), args.data());
+    const arg_vector args = {"test", "--verbosity", "high"};
+    auto input = snatch::cli::parse_arguments(static_cast<int>(args.size()), args.data());
 
     REQUIRE(input.has_value());
     CHECK(input->executable == "test"sv);
@@ -78,8 +78,8 @@ TEST_CASE("parse arguments verbosity (long form)", "[cli]") {
 };
 
 TEST_CASE("parse arguments verbosity (short form)", "[cli]") {
-    const arg_vector args  = {"test", "-v", "high"};
-    auto             input = snatch::cli::parse_arguments(args.size(), args.data());
+    const arg_vector args = {"test", "-v", "high"};
+    auto input = snatch::cli::parse_arguments(static_cast<int>(args.size()), args.data());
 
     REQUIRE(input.has_value());
     CHECK(input->executable == "test"sv);
@@ -92,15 +92,15 @@ TEST_CASE("parse arguments verbosity (short form)", "[cli]") {
 };
 
 TEST_CASE("parse arguments verbosity (no value)", "[cli]") {
-    const arg_vector args  = {"test", "--verbosity"};
-    auto             input = snatch::cli::parse_arguments(args.size(), args.data());
+    const arg_vector args = {"test", "--verbosity"};
+    auto input = snatch::cli::parse_arguments(static_cast<int>(args.size()), args.data());
 
     CHECK(!input.has_value());
 };
 
 TEST_CASE("parse arguments unknown", "[cli]") {
-    const arg_vector args  = {"test", "--make-coffee"};
-    auto             input = snatch::cli::parse_arguments(args.size(), args.data());
+    const arg_vector args = {"test", "--make-coffee"};
+    auto input = snatch::cli::parse_arguments(static_cast<int>(args.size()), args.data());
 
     REQUIRE(input.has_value());
     CHECK(input->executable == "test"sv);
@@ -108,8 +108,8 @@ TEST_CASE("parse arguments unknown", "[cli]") {
 };
 
 TEST_CASE("parse arguments positional", "[cli]") {
-    const arg_vector args  = {"test", "arg1"};
-    auto             input = snatch::cli::parse_arguments(args.size(), args.data());
+    const arg_vector args = {"test", "arg1"};
+    auto input = snatch::cli::parse_arguments(static_cast<int>(args.size()), args.data());
 
     REQUIRE(input.has_value());
     CHECK(input->executable == "test"sv);
@@ -122,8 +122,8 @@ TEST_CASE("parse arguments positional", "[cli]") {
 };
 
 TEST_CASE("parse arguments too many positional", "[cli]") {
-    const arg_vector args  = {"test", "arg1", "arg2"};
-    auto             input = snatch::cli::parse_arguments(args.size(), args.data());
+    const arg_vector args = {"test", "arg1", "arg2"};
+    auto input = snatch::cli::parse_arguments(static_cast<int>(args.size()), args.data());
 
     REQUIRE(!input.has_value());
 };
